feat : 파일이동 및 배경 적용을 위한 속성 추가
