--- pages/test.tsx
@@ -1,9 +1,0 @@
-import AnimatedCircles from "../features/collection/background-animation/BackgroundCircles";
-
-export default function Page() {
-  return (
-    <>
-      <AnimatedCircles />
-    </>
-  );
-}
